[#23] Implement backends for survey detail
